requirements.txt
django>=2.0
django_materialize_nav
